# priv/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Canary API",
    "version": "0.1.0",
    "description": "Machine-consumable contract for the Canary observability API. Canary exposes deterministic, bounded responses for error ingestion, health checks, incidents, timelines, operator management, and webhook delivery visibility.",
    "x-agent-guide": {
      "summary": "Register webhooks as wake-up hints, then treat the timeline as the durable source of truth for correctness.",
      "flow": [
        "Create a webhook subscription for the business events that should wake the agent.",
        "Persist the last processed timeline cursor after each successful polling cycle.",
        "When a webhook arrives, poll GET /api/v1/timeline from the last persisted cursor.",
        "Process the returned events in order, then persist the response cursor before going idle again.",
        "Write incident or error-group annotations back to Canary when the agent completes follow-up work."
      ],
...
      "crash_recovery": "After a crash or restart, resume from the last persisted cursor, replay missed timeline events, and continue polling until the API returns no further cursor.",
      "cursor_precedence": "When both `after` and `cursor` are provided on timeline-style pagination endpoints, Canary resolves `after` first and ignores `cursor` for that request.",
...
      "cold_start": {
...
      },
      "annotation_write_back": {
...
        "stable_actions": [
          "triaged",
...
        ],
...
        ],
...
  "servers": [
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/targets": {
      "get": {
        "tags": [
          "Targets"
        ],
        "summary": "List health-check targets",
        "responses": {
          "200": {
            "description": "Target list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TargetsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          }
        },
        "x-canary-required-scope": "admin"
      },
      "post": {
        "tags": [
          "Targets"
        ],
        "summary": "Create a health-check target",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Target created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Target"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "422": {
            "$ref": "#/components/responses/ValidationProblem"
          }
        },
        "x-canary-required-scope": "admin"
      }
    },
...
    "/api/v1/monitors": {
      "get": {
        "tags": [
          "Monitors"
        ],
        "summary": "List non-HTTP health monitors",
        "responses": {
          "200": {
            "description": "Monitor list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitorsResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          }
        },
        "x-canary-required-scope": "admin"
      },
      "post": {
        "tags": [
          "Monitors"
        ],
        "summary": "Create a non-HTTP health monitor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitorCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Monitor created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Monitor"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "422": {
            "$ref": "#/components/responses/ValidationProblem"
          }
        },
        "x-canary-required-scope": "admin"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Send `Authorization: Bearer sk_live_...` with an `ingest-only`, `read-only`, or `admin` API key."
      }
    },
...
    "responses": {
      "UnauthorizedProblem": {
        "description": "Missing, invalid, or revoked API key.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
      "ValidationProblem": {
        "description": "Request validation failed.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      },
...
      "ForbiddenProblem": {
        "description": "API key is valid but lacks the required scope.",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "request_id"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "request_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "retry_after": {
            "type": "integer"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "Target": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "service",
          "url",
          "method",
          "interval_ms",
          "timeout_ms",
          "expected_status",
          "active",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "HEAD"
            ]
          },
          "interval_ms": {
            "type": "integer"
          },
          "timeout_ms": {
            "type": "integer"
          },
          "expected_status": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TargetsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "targets"
        ],
        "properties": {
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Target"
            }
          }
        }
      },
      "TargetCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "url"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "method": {
            "type": "string",
            "enum": [
              "GET",
              "HEAD"
            ]
          },
          "headers": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "interval_ms": {
            "type": "integer",
            "minimum": 1000
          },
          "timeout_ms": {
            "type": "integer",
            "minimum": 1
          },
          "expected_status": {
            "type": "string"
          },
          "body_contains": {
            "type": "string"
          },
          "degraded_after": {
            "type": "integer"
          },
          "down_after": {
            "type": "integer"
          },
          "up_after": {
            "type": "integer"
          },
          "allow_private": {
            "type": "boolean",
            "description": "Compatibility flag. Private target URLs are accepted only when the server is configured with ALLOW_PRIVATE_TARGETS=true."
          }
        }
      },
...
      "Monitor": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "name",
          "service",
          "mode",
          "expected_every_ms",
          "grace_ms",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "schedule",
              "ttl"
            ]
          },
          "expected_every_ms": {
            "type": "integer",
            "minimum": 1
          },
          "grace_ms": {
            "type": "integer",
            "minimum": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MonitorsResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "monitors"
        ],
        "properties": {
          "monitors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Monitor"
            }
          }
        }
      },
      "MonitorCreateRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "mode",
          "expected_every_ms"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "service": {
            "type": "string"
          },
          "mode": {
            "type": "string",
            "enum": [
              "schedule",
              "ttl"
            ]
          },
          "expected_every_ms": {
            "type": "integer",
            "minimum": 1
          },
          "grace_ms": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
...
    }
  }
}